.Net: Add support for OpenAPI parameters defined with schema but without a type (#10682)
### Motivation, Context and Description  
Currently, the SK OpenAPI functionality supports operations with
parameters defined by a schema that has a `type` property. If this
property is not provided, which can happen if the schema is described
using the `anyOf`, `allOf`, or `oneOf` keywords, the execution of the
operation with such parameters fails.
   
This PR relaxes the OpenAPI parameters validation/conversion
functionality by allowing parameters with no `type` property that has a
valid schema or no schema at all.

Fixes the _"System.NotSupportedException: 'Unexpected type '' of
parameter 'lat' with argument '37.7749'.'"_ mentioned in this issue:
https://github.com/microsoft/semantic-kernel/issues/10548