.Net: scrub preceding cross region identifier in Amazon Bedrock model IDs (#10859)
…fix from model id

If present, removes the "us.", "eu.", or "apac." preceding the model id
when selecting the correct model service for the given model.

### Motivation and Context

1. Why is this change required? This is to allow the use of cross-region
inference models in the Amazon Bedrock connector.
2. What problem does it solve? This fixes a bug in
`BedrockServiceFactory.cs` in which the cross-regional identifier in a
model was getting treated as an LLM provider.
3. What scenario does it contribute to? Use of cross-regional inference
models in Amazon Bedrock, namely Claude 3.7 Sonnet from a us-west-2 data
center.
4. If it fixes an open issue, please link to the issue here. 
   * #10738 
   * #10047 

### Description

This is a minimal-change quick fix for enabling cross-regional models in
Amazon Bedrock. It looks for one of the three standard cross-regional
identifiers at the start of the model ID in the bedrock service factory,
and if present, removes it. With the model ID stripped of the
cross-region header, existing logic is used to select the correct LLM
service.

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [X] The code builds clean without any errors or warnings
- [X] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [X] All unit tests pass, and I have added new tests where possible
- [X] I didn't break anyone :smile:

---------

Co-authored-by: Roger Barreto <19890735+RogerBarreto@users.noreply.github.com>