.Net MEVD: SqlServerVectorStore should accept a connection string to be thread safe (#11042)
By accepting a connection string rather than a connection itself we can
allow the store to be thread safe and we don't need to worry about the
lifetime of provided connection.

It's a breaking change, but we really want to steer the users to do the
right thing, so the sooner we release it, the better.