.Net: Add cancellation token & add headers to request (#11900)
### Motivation, Context and Description

This PR:
1. Propagates the cancellation token from all kernel functions of the
`SessionsPythonPlugin` plugin down to the APIs they call.
2. Refactors the way the plugin adds headers to the request. Instead of
adding them as default headers of the HTTP client, they are now added as
request headers. The result is the same - the headers are sent each call
of any kernel function of the plugin. This change may save time later by
preventing bug troubleshooting if/when an HTTP client is injected into
the plugin and it sets default headers to the client it does not own.
3. Replaces the usage of HTTP client methods for getting content, such
as `ReadAsStringAsync`, with the SK wrappers
`ReadAsStringWithExceptionMappingAsync` to align the plugin's behavior
with that of other SK components.

Contributes to:
https://github.com/microsoft/semantic-kernel/issues/10070