.Net: Stop setting dimensions on MEVD calls to IEmbeddingGenerator (#11908)
The current implementation of MEVD always passes the user-configured
dimensions on the vector property to the embedding generator
([code](https://github.com/microsoft/semantic-kernel/blob/feature-vector-data-preb3/dotnet/src/Connectors/VectorData.Abstractions/ProviderServices/VectorPropertyModel.cs#L129)).
Unfortunately, Weslie found out that the Azure OpenAI embedding
generator apparently errors when dimensions are specified (regardless
what non-null value is specified). Note that apparently the non-Azure
OpenAI service doesn't error (the value is simply ignored) - just the
Azure one.

To make matters slightly worse, apparently some models **do** allow
specifying dimensions, while others (on the same service) do not. This
means that aside from starting to hard code which models allow and which
don't in the embedding generator (which we probably don't want to do),
this decision must be offloaded to the user.

This PR removes the setting of dimensions on the MEVD calls to the
IEmbeddingGenerator. MEAI IEmbeddingGenerators which support varying
dimensions already allow for the default dimensions to be specified at
construction time. Although that would mean users need to specify them
twice (once for the default when constructing the embedding generator,
once for MEVD for e.g. collection creation), it seems to be the most
flexible option.

Another option would be modify the Azure OpenAI generator to allow
opting out from specifying the dimensions, though that will be quite
undiscoverable.

@stephentoub @SteveSandersonMS any thoughts here?

/cc @westey-m

Co-authored-by: Mark Wallace <127216156+markwallace-microsoft@users.noreply.github.com>