.Net: Use type to represent python code execution result (#11931)
### Motivation, Context and Description

This PR introduces a new type, `SessionsPythonCodeExecutionResult`, to
represent the Python code execution result and refactors the
`SessionsPythonPlugin.ExecuteCodeAsync` method to return this type
instead of a string. This should allow access to the result in a
structured way for scenarios where the result needs to be processed
further, for example, in the auto function invocation filters. At the
same time, the type overloads the `ToString()` method to be used in
scenarios where the result needs to be a string, for example when the
function is invoked as part of rendering a prompt template.

Contributes to:
https://github.com/microsoft/semantic-kernel/issues/10070