.Net: Add format script helper (#11905)
### Motivation and Context

As a small improvement to the C# development life cycle, this scripts
allows us to identify and correct quickly on branches any format warning
that we might have missed.

This will format only the projects that had code changes during the
development of your current branch compared to main, ignoring the ones
that were untouched.

This was inspired in the current logic that exists in the pipeline, but
could easily be ran locally for simplification and faster development.