.Net: Remove the HTTPS validations in AzureClientCore (#11943)
### Remove HTTPS validation from `AzureClientCore` class.

We are redirecting requests internally to inject API keys securely and
keep track of token usage in a transparent way, but this involves a YARP
proxy in between certain applications and Azure OpenAI. That proxy,
being internal behind our servers, doesn't use HTTPS and it doesn't need
to use it either. The current version of Semantic Kernel throws an error
when using a HTTP endpoint. Current fixes involve a lot of additional
steps and going out of the server to come back in through our azure
gateway or adding self-generated certificates and bypassing SSL errors,
etc... This fix would be very helpful.

* Remove the `Verify.StartsWith` validation for the `endpoint` parameter
in the constructors of `AzureClientCore` class.
* Remove the test case
`ConstructorWithInvalidEndpointShouldThrowArgumentException` from
`AzureClientCoreTests` class.
* Ensure the remaining test cases do not enforce HTTPS validation.