.Net: Remove SqlDataReaderDictionary in SQL Server provider (#12241)
We currently write records to the database by first serializing them to
a Dictionary, and then writing that Dictionary out as SqlParameters on
the command. Similarly, we read records by allocating and populating a
Dictionary, only to then iterate over that Dictionary and populate the
.NET instance from it.

This PR does away with the intermediate Dictionary representation. This
optimizes serialization by removing the double-copy, reduces
allocations, opens the way for boxing-free serialization, and also just
simplifies the code (less conversions and stuff to follow).