.Net: Add Usage Metadata for ChatClientChatCompletionService Adapter (#12412)
### Motivation and Context

As part of Agents code-base is commong practice now to acquire `Usage`
from the ChatMessageContent.Metadata, although is different how this is
done with the new `Microsoft.Extensions.AI` relying on the
`UsageDetails` or `UsageContent` to avoid any disruption when migrating
`Agent`s to use the `IChatClient`. This change was made, so old code can
still function.

Co-authored-by: Mark Wallace <127216156+markwallace-microsoft@users.noreply.github.com>