Python: added operationid validation in openapi spec (#12551)
Description
This PR enforces the presence of an explicit `operationId` for every
operation in the OpenAPI specification. Previously, if `operationId` was
missing, a default value was generated using the path and HTTP method.
Now, the parser will raise a PluginInitializationError if operationId is
not provided, ensuring that all operations are uniquely and explicitly
identified as required by the OpenAPI standard.

Key changes:

- Added validation to check for the existence of operationId in each
operation.
- If operationId is missing, a clear error is raised, preventing
fallback to a default value.
-
[python/tests/unit/connectors/openapi_plugin/no-operationid-openapi.yaml](https://github.com/microsoft/semantic-kernel/compare/main...sshandilya1991:semantic-kernel:sshandilya/12443_add_operation_id_validation?expand=1#diff-3d272039b3afa97f89805f0ee0258a98b613f0f373c603467b336a8912d5651a)
serves as a test case.
- Closes #12443

Edit:
- Added more validation for operationId to be unique & throw in case
duplicate operationId found while parsing.