.NET: Update dependabot schedule (#12602)
### Motivation, Context and Description

This PR updates the dependabot schedule to run twice a week instead of
weekly, with one of the days being Sunday. The idea is that the
dependabot infrastructure is less busy on weekends. This change should
decrease the number of dependencies to update and potentially resolve
the dependabot timeout issue. Additionally, the schedule for checking
dependencies of github actions has been changed to Tuesday to avoid
running on Monday, when the job for checking dependencies for samples is
executed.

Dependabot options -
https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference