.Net: Refactor kernel initialization in Step01_Agent (#12181)
This pull request updates the `UseSingleChatCompletionAgent` method in
`Step01_Agent.cs` to improve code clarity and reusability. The most
important change is replacing the call to
`CreateKernelWithChatCompletion()` with the existing `kernel` parameter.

Code clarity and reusability:

*
[`dotnet/samples/GettingStartedWithAgents/Step01_Agent.cs`](diffhunk://#diff-406533ddffa54164a374596ac14a8eb4f1b07ed75f8a235d349c4086cc69dbc5L36-R36):
Updated the `Kernel` property to use the `kernel` parameter instead of
calling `CreateKernelWithChatCompletion()`. This change removes
redundancy and simplifies the method.

Co-authored-by: Chris <66376200+crickman@users.noreply.github.com>
Co-authored-by: Mark Wallace <127216156+markwallace-microsoft@users.noreply.github.com>